cbox module renamed
